Start of settings object
